+++ network/Sources/main/Broadcast.swift
@@ -3,9 +3,9 @@ import Network
 
 class Broadcast {
     let host: String
-    let port: UInt16
+    let port: Int
 
-    init(at host: String, port: UInt16) {
+    init(at host: String, port: Int) {
         self.host = host
         self.port = port
     }
